fix: cast BIGINT to INTEGER in XP calculation + helper script for email
PROBLEM:
- 1-FIX-XP-SAFE.sql: calculate_level(bigint) error
- 2-FIX-RLS-ADMIN-SAFE.sql: User doesn't know which email to use

SOLUTION:
- Added ::INTEGER cast to SUM() result (line 161-162)
- SUM() returns BIGINT but calculate_level() expects INTEGER
- Created TROVA-TUA-EMAIL.sql helper script

TROVA-TUA-EMAIL.sql:
- Shows all users with email and UUID
- User can find their own email
- Then copy UUID and set admin role
- Simpler than editing email in script

+++ 1-FIX-XP-SAFE.sql
@@ -158,8 +158,8 @@ INSERT INTO user_progress (
 )
 SELECT 
   qr.user_id,
-  COALESCE(SUM(qr.correct_answers * 10), 0) as total_xp,
-  calculate_level(COALESCE(SUM(qr.correct_answers * 10), 0)) as level,
+  COALESCE(SUM(qr.correct_answers * 10), 0)::INTEGER as total_xp,
+  calculate_level(COALESCE(SUM(qr.correct_answers * 10), 0)::INTEGER) as level,
   COUNT(qr.id) as total_quizzes,
   COALESCE(SUM(qr.total_questions), 0) as total_q,
   COALESCE(SUM(qr.correct_answers), 0) as correct,
